Move test files into its own file
